fix-cloudinary-images

+++ app/Http/Controllers/ProfileController.php
@@ -32,17 +32,23 @@ class ProfileController extends Controller
      */
     public function update(ProfileRequest $request, Profile $profile)
     {
-        //$this->authorize('update', $profile);
+        $this->authorize('update', $profile);
         $user = Auth::user();
-
         $current_photo = $profile->photo;
-        $split_url = explode('/', $current_photo);
-        $public_id = explode('.', $split_url[sizeof($split_url)-1]);
+        
+        if ($current_photo) {
+            $split_url = explode('/', $current_photo);
+            $public_id = explode('.', end($split_url))[0];
+        } else {
+            $public_id = null; // O algún valor por defecto si no hay foto
+        }
+
+        $photo = $current_photo;
 
         if($request->hasFile('photo')){
             Cloudinary::destroy('profiles/' . $public_id[0]);
             // Set new photo
-            $article['photo'] = Cloudinary::upload($request->file('image')
+            $photo = Cloudinary::upload($request->file('photo')
             ->getRealPath(),[
                 'folder' => 'profiles',
             ])->getSecurePath();
@@ -51,7 +57,7 @@ class ProfileController extends Controller
             $photo = $user->profile->photo;
         }
 
-        // Set nae, email and photo
+        // Set name, email and photo
         $user->name = $request->name;
         $user->email = $request->email;
         $user->profile->photo = $photo;
